Organize auth files into a protectedroute
